refactor: move gson to stdwrap
