+++ README.md
@@ -552,7 +552,7 @@ Usage：
 
 ```go
 import (
-    "github.com/bytedance/gg/gson"
+    "github.com/bytedance/gg/stdwrap/gson"
 )
 ```
