+++ collection/skipmap/skipmap_example_test.go
@@ -19,7 +19,7 @@ import (
 	"strconv"
 	"sync"
 
-	"github.com/bytedance/gg/gson"
+	"github.com/bytedance/gg/stdwrap/gson"
 )
 
 func Example() {
