+++ gslice/gslice_example_test.go
@@ -18,8 +18,8 @@ import (
 	"fmt"
 	"strconv"
 
-	"github.com/bytedance/gg/gson"
 	"github.com/bytedance/gg/gvalue"
+	"github.com/bytedance/gg/stdwrap/gson"
 )
 
 func Example() {
